# docs/swagger.yaml
basePath: /v0
definitions:
...
  inputs.ProjectAlertGroupBody:
    properties:
      name:
        type: string
      project_id:
        type: integer
      service:
        type: string
    required:
    - name
    - project_id
    - service
    type: object
...
  models.Response:
    properties:
      data:
        type: object
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: I have no specific description
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hod Hod API
  version: "1.0"
paths:
...
  /alert-group/{id}/project:
    get:
      description: return ProjectsAlertGroup
      operationId: get_projects_of_AlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get ProjectsAlertGroup
      tags:
      - Project Alert Group
    post:
      description: new ProjectAlertGroup
      operationId: new_ProjectAlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/inputs.ProjectAlertGroupBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: new ProjectAlertGroup
      tags:
      - Project Alert Group
  /alert-group/{id}/project/{pid}:
    delete:
      description: delete ProjectAlertGroup by ProjectAlertGroupID
      operationId: delete_project_alert_group
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group_id
        in: path
        name: pid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: delete ProjectAlertGroup
      tags:
      - Project Alert Group
    get:
      description: return ProjectAlertGroup
      operationId: get_ProjectAlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_alert_group_id
        in: path
        name: pid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get ProjectAlertGroup
      tags:
      - Project Alert Group
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"